# src/main/resources/api/openapi.yaml
openapi: 3.1.1
info:
  title: QozzBank API
  description: >
    Core Banking System API.
    All endpoints require a valid JWT Bearer token from Supabase Auth.
    User identification is handled automatically server-side via the JWT context.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

security:
  - BearerAuth: [ ]

tags:
  - name: users
    description: Current user profile management
  - name: accounts
    description: Bank account management and overview
  - name: transactions
    description: Money transfers and historical records

paths:
  /:
    get:
      tags: [ home ]
      security: [ ]
      summary: Service Summary
      operationId: getHomeInfo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeInfoResponse' }

  /api/v1/users/me:
    get:
      tags: [ users ]
      summary: Get current user details
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { description: "User record not found" }

  /api/v1/users:
    post:
      tags: [ users ]
      summary: Register new user
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRequest' }
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }

  /api/v1/accounts:
    get:
      tags: [ accounts ]
      summary: List my accounts
      operationId: getAccounts
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: List of accounts retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountsResponse' }
    post:
      tags: [ accounts ]
      summary: Open a new account
      operationId: createAccount
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountRequest' }
      responses:
        '201':
          description: Account opened successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }

  /api/v1/accounts/{iban}:
    get:
      tags: [ accounts ]
      summary: Get account details by IBAN
      operationId: getAccount
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IbanPath'
      responses:
        '200':
          description: Account data found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '404': { description: "Account not found" }

  /api/v1/accounts/{iban}/transactions:
    get:
      tags: [ accounts ]
      summary: Get transaction history for an account
      operationId: getAccountTransactions
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IbanPath'
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionsResponse' }

  /api/v1/transactions:
    post:
      tags: [ transactions ]
      summary: Execute money transfer
      operationId: createTransaction
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountsTransactionRequest' }
      responses:
        '201':
          description: Transfer accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '422':
          description: Business validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    UnauthorizedError:
      description: Authentication failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

  parameters:
    IbanPath:
      name: iban
      in: path
      required: true
      schema: { type: string, example: "USD12345678901234567890" }
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }

  schemas:
    # --- WRAPPERS ---
    UserResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/UserData' }

    AccountsResponse:
      type: object
      properties:
        accounts:
          type: array
          items: { $ref: '#/components/schemas/AccountData' }
        totalCount: { type: integer, example: 2 }

    AccountResponse:
      type: object
      properties:
        account: { $ref: '#/components/schemas/AccountData' }

    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items: { $ref: '#/components/schemas/TransactionData' }
        totalCount: { type: integer, example: 15 }

    TransactionResponse:
      type: object
      properties:
        transaction: { $ref: '#/components/schemas/TransactionData' }

    # --- MODELS ---
    HomeInfoResponse:
      type: object
      properties:
        message: { type: string, example: "Welcome to QozzBank API" }

    UserRequest:
      type: object
      required: [ firstName, lastName, email ]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }

    UserData:
      type: object
      required: [ firstName, lastName, email, status, createdAt ]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        status: { type: string, example: "ACTIVE" }
        createdAt: { type: string, format: date-time }

    AccountRequest:
      type: object
      required: [ accountType, currency ]
      properties:
        accountType: { type: string, enum: [ Checking, Savings, Deposit, Credit, Investment, Loan ] }
        currency: { type: string, example: "USD" }

    AccountData:
      type: object
      required: [ iban, currency, balance, availableBalance, accountType, status, createdAt ]
      properties:
        iban: { type: string }
        currency: { type: string }
        balance: { type: number }
        availableBalance: { type: number }
        accountType: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }

    TransactionData:
      type: object
      required: [ id, operationId, amount, type, status, createdAt ]
      properties:
        id: { type: string, format: uuid }
        operationId: { type: string, format: uuid }
        fromIban: { type: string, example: "USD1234567890..." }
        toIban: { type: string, example: "EUR9876543210..." }
        amount: { type: number }
        type: { type: string, example: "DEBIT" }
        status: { type: string, example: "COMPLETED" }
        description: { type: string }
        createdAt: { type: string, format: date-time }

    AccountsTransactionRequest:
      type: object
      required: [ fromIban, toIban, amount ]
      properties:
        fromIban: { type: string }
        toIban: { type: string }
        amount: { type: number, minimum: 0.01 }
        description: { type: string, maxLength: 255 }

    ErrorResponse:
      type: object
      required: [ code, message, correlationId ]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, example: "Insufficient funds" }
        correlationId: { type: string, format: uuid }

    ErrorCode:
      type: string
      enum:
        - bad_request.missing_header
        - bad_request.media_type_not_supported
        - unauthorized.invalid_token
        - business.insufficient_funds